_templates/strategy_template/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
